multi-draw, hit caps
I think this fixes the multi-draw issues so it works now. Still need to actually add tests, but it should support having the _pull() tested now.

Added hit cap, miss cap, sum cap, sum floor, and a flag for whether to only hold to the hit cap if it can crit

Added Config parameters to the bagpool.conf. A default level for all bags, and a bag specific level. These can also be configured at run time, but the command line arguments for it haven't been created yet.